moved AuctionTimer, BidAmount and HighestBid to /modules/common/corgi/auction
