+++ src/modules/common/corgi/auction/BidAmount/BidAmount.scss
