+++ src/modules/common/corgi/auction/BidAmount/index.js
