+++ src/modules/common/corgi/auction/HighestBid/HighestBid.scss
