+++ src/modules/common/corgi/auction/HighestBid/index.js
