+++ src/modules/corgi/components/BidTile/BidTile.jsx
@@ -5,7 +5,7 @@ import './BidTile.scss';
 import { NearContext } from '~contexts';
 
 import { Owner } from '~modules/common';
-import { BidAmount } from '~modules/corgi/components';
+import { BidAmount } from '~modules/common/corgi';
 
 import { humanizeTime } from '~helpers/time';
