+++ src/modules/corgi/components/index.js
@@ -1,10 +1,7 @@
 import AuctionCard from './AuctionCard';
-import AuctionTimer from './AuctionTimer';
-import BidAmount from './BidAmount';
-import BidHistory from './BidHistory';
 import BidDifference from './BidDifference';
+import BidHistory from './BidHistory';
 import BidTile from './BidTile';
-import HighestBid from './HighestBid';
 import TimeTile from './TimeTile';
 
-export { AuctionCard, AuctionTimer, BidAmount, BidDifference, BidHistory, BidTile, HighestBid, TimeTile };
+export { AuctionCard, BidDifference, BidHistory, BidTile, TimeTile };
